+++ src/main/java/com/hyeon/guardrail/file/domain/FileTargetType.java
@@ -1,0 +1,6 @@
+package com.hyeon.guardrail.file.domain;
+
+/** 파일 첨부 대상 타입 */
+public enum FileTargetType {
+  PRODUCT
+}